refactor: start refactoring styles
