+++ packages/key-facts/README.md
@@ -1,7 +1,7 @@
 # KeyFacts
 
-Unordered list of textual AST information that can contain text, links and styled
-text such as bold, or italic content.
+Unordered list of textual AST information that can contain text, links and
+styled text such as bold, or italic content.
 
 ## Contributing
