+++ packages/key-facts/__tests__/shared-key-facts.native.js
@@ -1,4 +1,3 @@
-import React from "react";
 import { shallow } from "enzyme";
 import {
   addSerializers,
@@ -10,11 +9,7 @@ import {
   propsNoChildren,
   replaceTransform
 } from "@times-components/jest-serializer";
-import { iterator } from "@times-components/test-utils";
-import KeyFacts from "../src/key-facts";
-import data from "../fixtures/key-facts-tests.json";
-
-const { data: { children, attributes } } = data;
+import renderKeyFacts from "./shared-render-key-facts";
 
 export default () => {
   addSerializers(
@@ -30,32 +25,5 @@ export default () => {
     )
   );
 
-  const tests = [
-    {
-      name: "key facts with title",
-      test: () => {
-        const wrapper = shallow(
-          <KeyFacts
-            items={children[0].children}
-            onLinkPress={() => {}}
-            title={attributes.title}
-          />
-        );
-
-        expect(wrapper).toMatchSnapshot();
-      }
-    },
-    {
-      name: "key facts without title",
-      test: () => {
-        const wrapper = shallow(
-          <KeyFacts items={children[0].children} onLinkPress={() => {}} />
-        );
-
-        expect(wrapper).toMatchSnapshot();
-      }
-    }
-  ];
-
-  iterator(tests);
+  renderKeyFacts(shallow);
 };
