+++ packages/key-facts/package.json
@@ -46,11 +46,13 @@
     "enzyme": "3.3.0",
     "eslint": "4.9.0",
     "jest": "21.2.1",
+    "jest-styled-components": "5.0.0",
     "prettier": "1.8.2",
     "react": "16.4.1",
     "react-dom": "16.3.1",
     "react-native": "0.54.2",
     "react-test-renderer": "16.3.1",
+    "snapshot-diff": "0.4.0",
     "webpack": "4.6.0"
   },
   "dependencies": {
